+++ CHANGELOG.md
@@ -7,6 +7,26 @@
 - Change behavior of `TraverseConfig::traverse` to continuously start up new `Indexer::on_finalized` tasks up to `max_parallel` as the output channel takes the results of prior tasks.
   The prior behavior was to start preparing `max_parallel` tasks, then wait for everything to be consumed before starting another round of `max_parallel` tasks.
 - Introduce `RewardsOverview::common_reward_data` for getting the common reward data across `RewardsOverview` version 0 and version 1.
+- Add functionality for generating, and verifying account signatures.
+- Support for protocol version 8 functionality:
+  - `ConfigureBakerPayload` supports the optional `suspend` flag.
+  - `BakerEvent` has new cases for `BakerSuspended` and `BakerResumed` when the flag is set in a
+    `ConfigureBaker` transaction.
+  - `SpecialTransactionOutcome` has new cases for `ValidatorSuspended` and
+    `ValidatorPrimedForSuspension`, which occur when a validator is (or potentially will be)
+    suspended.
+  - New `UpdatePayload` type `ValidatorScoreParametersCPV3`, which updates the maximum number of
+    consecutive failures a validator can have before it faces suspension.
+- `ContractInitializedEvent` adds the `parameter` used to initialize the contract (supported from
+  node version >= 8).
+- New functionality for querying which accounts have scheduled releases or cooldowns (supported
+  from node version >= 8):
+  - `get_scheduled_release_accounts`: Get the accounts (by index) with scheduled releases, and the
+    timestamp of the first release.
+  - `get_cooldown_accounts`: Get the accounts (by index) with stake in cooldown, and the timestamp
+    at which the first cooldown expires.
+  - `get_pre_cooldown_accounts`: Get the accounts (by index) with stake in pre-cooldown.
+  - `get_pre_pre_cooldown_accounts`: Get the accounts (by index) with stake in pre-pre-cooldown.
 
 ## 5.0.0